requirements-desktop.txt
-r requirements.txt
pywebview>=5.3,<6.0
pyinstaller>=6.10,<7.0
